Backend/api/requirements.txt
adversarial-robustness-toolbox==1.16.0
gcloud==0.18.3
fastapi==0.100.0
uvicorn==0.23.1
func-timeout==4.3.5
google-cloud-pubsub==2.18.4
google-cloud-firestore==2.14.0
google-cloud-storage
google-cloud-container
torch==1.13.1
cma
torchvision==0.14.1
python-dotenv==1.0.0
dill==0.3.8
numpy==1.24.3
keras==2.13.1
tensorflow==2.13.0
tensorboard==2.13.0
tensorflow-datasets==4.9.3
tensorflow-estimator==2.13.0
tensorflow-metadata==1.14.0
xgboost==2.0.3
pandas
scikit-learn
catboost==1.2.2
scikit-optimize==0.9.0
pdfkit
optuna
category_encoders
ctgan
xhtml2pdf
seaborn
lightgbm
PyPDF2
pytz
asyncio
fpdf2
kubernetes
TTTS
google-cloud-secret-manager
